Camera y axis was flipped
Discovered while working on next step of the model and getting very confusing
results. Spent time debugging the ray-sphere intersection code under false
assumption that code was broken for cases where the ray origin is inside the
sphere.

Fixing the camera issue also highlighted that UV coordinate system generation
was not to my liking (texture was flipped vertically and horizontally). The
vertical flip was masked by the upside down camera (an even number of sign
errors cancel out) and the horizontal flip was obvious in hindsight (visible
in worklog images with Florida and California flipped in the image). Oh well.